+++ sdk/appconfiguration/azure-appconfiguration/azure/appconfiguration/_models.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License.
 # ------------------------------------
 import json
-from typing import Any, Union
+from typing import Any, Union, List, Dict, Optional
 from ._generated._serialization import Model
 from ._generated.models import KeyValue
 
@@ -50,7 +50,7 @@ class ConfigurationSetting(Model):
     kind = "Generic"
     content_type = None
 
-    def __init__(self, **kwargs: Any) -> None:
+    def __init__(self, **kwargs) -> None:
         super(ConfigurationSetting, self).__init__(**kwargs)
         self.key = kwargs.get("key", None)
         self.label = kwargs.get("label", None)
@@ -155,7 +155,14 @@ class FeatureFlagConfigurationSetting(ConfigurationSetting): # pylint: disable=t
     )
     kind = "FeatureFlag"
 
-    def __init__(self, feature_id: str, **kwargs: Any) -> None:  # pylint: disable=dangerous-default-value, super-init-not-called
+    def __init__( # pylint: disable=super-init-not-called
+        self,
+        feature_id: str,
+        *,
+        enabled: Optional[bool] = None,
+        filters: Optional[List[Dict[str, Any]]] = None,
+        **kwargs
+    ) -> None:
         if "key" in kwargs.keys() or "value" in kwargs.keys():
             raise TypeError("Unexpected keyword argument, do not provide 'key' or 'value' as a keyword-arg")
         self.feature_id = feature_id
@@ -168,8 +175,8 @@ class FeatureFlagConfigurationSetting(ConfigurationSetting): # pylint: disable=t
         self.etag = kwargs.get("etag", None)
         self.description = kwargs.get("description", None)
         self.display_name = kwargs.get("display_name", None)
-        self.filters = kwargs.get("filters", [])
-        self.enabled = kwargs.get("enabled", None)
+        self.filters = [] if filters is None else filters
+        self.enabled = enabled
         self._value = json.dumps({"enabled": self.enabled, "conditions": {"client_filters": self.filters}})
 
     @property
@@ -281,7 +288,7 @@ class SecretReferenceConfigurationSetting(ConfigurationSetting):
     )
     kind = "SecretReference"
 
-    def __init__(self, key: str, secret_id: str, **kwargs: Any) -> None: # pylint: disable=super-init-not-called
+    def __init__(self, key: str, secret_id: str, **kwargs) -> None: # pylint: disable=super-init-not-called
         if "value" in kwargs.keys():
             raise TypeError("Unexpected keyword argument, do not provide 'value' as a keyword-arg")
         self.key = key