+++ sdk/kubernetesconfiguration/azure-mgmt-kubernetesconfiguration/azure/mgmt/kubernetesconfiguration/v2021_05_01_preview/aio/operations/__init__.py
@@ -15,13 +15,19 @@ from ._location_extension_types_operations import LocationExtensionTypesOperatio
 from ._source_control_configurations_operations import SourceControlConfigurationsOperations
 from ._operations import Operations
 
+from ._patch import __all__ as _patch_all
+from ._patch import *  # pylint: disable=unused-wildcard-import
+from ._patch import patch_sdk as _patch_sdk
+
 __all__ = [
-    'ExtensionsOperations',
-    'OperationStatusOperations',
-    'ClusterExtensionTypeOperations',
-    'ClusterExtensionTypesOperations',
-    'ExtensionTypeVersionsOperations',
-    'LocationExtensionTypesOperations',
-    'SourceControlConfigurationsOperations',
-    'Operations',
+    "ExtensionsOperations",
+    "OperationStatusOperations",
+    "ClusterExtensionTypeOperations",
+    "ClusterExtensionTypesOperations",
+    "ExtensionTypeVersionsOperations",
+    "LocationExtensionTypesOperations",
+    "SourceControlConfigurationsOperations",
+    "Operations",
 ]
+__all__.extend([p for p in _patch_all if p not in __all__])
+_patch_sdk()