+++ sdk/textanalytics/azure-ai-textanalytics/samples/sample_model_version.py
@@ -67,7 +67,7 @@ def sample_model_version() -> None:
                 print(f"......Entity '{entity.text}' has category '{entity.category}'")
         elif action_result.is_error is True:
             print("......Is an error with code '{}' and message '{}'".format(
-                action_result.code, action_result.message
+                action_result.error.code, action_result.error.message
             ))