+++ sdk/textanalytics/azure-ai-textanalytics/samples/sample_multi_label_classify.py
@@ -72,7 +72,7 @@ def sample_classify_document_multi_label() -> None:
                 ))
         elif classification_result.is_error is True:
             print("Movie plot '{}' has an error with code '{}' and message '{}'".format(
-                doc, classification_result.code, classification_result.message
+                doc, classification_result.error.code, classification_result.error.message
             ))
     # [END multi_label_classify]