+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Release History
 
-## 1.4.0 (Unreleased)
+## 1.5.0 (Unreleased)
+
+### Features Added
+
+### Bugs Fixed
+
+### Other Changes
+
+- Added dependency on `azure-mgmt-resource`
+- Added dependency on `azure-mgmt-resourcegraph`
+
+## 1.4.0 (2023-02-07)
 
 ### Features Added
 - Added dedicated classes for each type of job service and updated the docstrings. The classes added are `JupyterLabJobService, SshJobService, TensorBoardJobService, VsCodeJobService` with a few properties specific to the type.